streamlit_app/requirements.txt
streamlit
pandas
matplotlib
dotenv
plotly.express
